+++ Sources/MercadoPagoCheckout/Views/CardFormScreen.swift
@@ -9,61 +9,6 @@ import MPComponents
 import CoreMethods
 
 struct CardFormScreen: View {
-    
-    struct CardFormData {
-        @CardFormValidate(
-            RequiredRule(
-                MPStrings.CardForm.CardNumber.errorEmpty
-            ),
-            CardNumberRule()
-        )
-        var cardNumber: String = ""
-        
-        @CardFormValidate(
-            RequiredRule( MPStrings.CardForm.CardHolder.errorEmpty
-            ),
-            CardHolderRule()
-        )
-        var cardHolder: String = ""
-        
-        @CardFormValidate(
-            RequiredRule( MPStrings.CardForm.Expiration.errorEmpty
-            ),
-            ExpirationDateRule()
-        )
-        var expirationDate: String = ""
-        
-        @CardFormValidate(
-            RequiredRule(MPStrings.CardForm.CVV.errorEmpty),
-            SecurityCodeRule()
-        )
-        var securityCode: String = ""
-        
-        @CardFormValidate(
-            RequiredRule(
-                MPStrings.CardForm.Document.errorEmpty
-            ),
-            DocumentRule()
-        )
-        var documentHolder: String = ""
-        
-        mutating func setSecurityCodeLength(_ length: Int) {
-            _securityCode.update(.securityCodeLength(length))
-        }
-
-        mutating func setDocumentLength(_ length: Int) {
-            _documentHolder.update(.documentLength(length))
-        }
-        
-        var isFormValid: Bool {
-            return _cardNumber.errorMessages.isEmpty
-            && _cardHolder.errorMessages.isEmpty
-            && _expirationDate.errorMessages.isEmpty
-            && _securityCode.errorMessages.isEmpty
-            && _documentHolder.errorMessages.isEmpty
-        }
-    }
-    
     @ObservedObject private var viewModel: CardFormViewModel
     
     // MARK: States View
